+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # TDE - Geo Address Changelog
 
-All notable changes to this project will be documented in this file.
+## 1.0.13 - 2021-09-29
+### Added
+- Implemented `PreviewableFieldInterface` to add an address field to the element index
 
 ## 1.0.1 - 2018-01-19
-### Fixes
-- Fixed mixing up address-fields as object or array.
+### Fixed
+- Fixed mixing up address-fields as object or array
 
 ## 1.0.0 - 2018-01-19
 ### Added